# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Biome Training Intelligence API
  version: 1.0.0
  description: |
    # Biome API Specification

    This API serves as the communication backbone between the Biome Dashboard (Next.js) and the AI Coaching Engine (FastAPI).
    It is designed with strict contract-driven principles, ensuring that all AI-generated content (Weekly Plans, Findings)
    adheres to predefined JSON schemas.

    ## Key Domains:
    - **Metrics:** Aggregated training statistics derived from DuckDB analytics.
    - **Plan:** AI orchestration for generating and revising personalized workout protocols.
    - **Memory:** Long-term persistence of coach-user interactions and reflections.
    - **Agent:** Real-time, stateful chat sessions with specialized AI coaching personas.

    All endpoints return standard HTTP status codes and leverage Pydantic for request/response validation.

paths:
  /metrics/overview:
    get:
      summary: Get high-level weekly KPIs
      description: Returns frequency, volume load, and weak point counts for the current training week.
      tags: [Metrics]
      responses:
        '200':
          description: Successful retrieval of weekly metrics.
          content:
            application/json:
              schema:
                type: object
                properties:
                  weekly_frequency: { type: integer }
                  total_volume_load_current_week: { type: number }
                  active_weak_points_count: { type: integer }

  /metrics/trends:
    get:
      summary: Get historical time-series data
      description: Returns a list of TrendPoints for charting. Supports filtering by exercise.
      tags: [Metrics]
      parameters:
        - name: metric
          in: query
          required: true
          schema:
            type: string
            enum: [volume_load, average_rpe, max_weight, weekly_frequency]
        - name: exercise
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Trend data points.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    date: { type: string, format: date }
                    value: { type: number }

  /plan/propose:
    post:
      summary: Generate a new weekly plan
      description: Triggers the multi-agent system to analyze history and draft a fresh protocol.
      tags: [Plan]
      responses:
        '200':
          description: A structured weekly training protocol.
          content:
            application/json:
              schema:
                $ref: './schemas/WeeklyPlan.schema.json'

  /plan/revise:
    post:
      summary: Update a plan via feedback
      description: Takes a current plan and natural language feedback to produce a revised version.
      tags: [Plan]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_plan:
                  $ref: './schemas/WeeklyPlan.schema.json'
                feedback:
                  type: string
      responses:
        '200':
          description: The updated weekly training protocol.
          content:
            application/json:
              schema:
                $ref: './schemas/WeeklyPlan.schema.json'

  /plan/validate:
    post:
      summary: Audit a plan for safety
      description: Checks a protocol against safety constraints (e.g. excessive volume).
      tags: [Plan]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas/WeeklyPlan.schema.json'
      responses:
        '200':
          description: Validation result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: { type: boolean }
                  issues:
                    type: array
                    items: { type: string }

  /memory/write:
    post:
      summary: Persist a coaching signal
      description: Stores a summary of findings or plans in the long-term memory store.
      tags: [Memory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas/MemoryRecord.schema.json'
      responses:
        '200':
          description: Confirmation of record persistence.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  status: { type: string }

  /memory/search:
    post:
      summary: Search historical signals
      description: Filters memory records by keyword or snapshot type.
      tags: [Memory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query: { type: string }
                type: { type: string }
                limit: { type: integer }
      responses:
        '200':
          description: Matching memory records.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './schemas/MemoryRecord.schema.json'

  /memory/timeline:
    get:
      summary: Get recent signal stream
      description: Returns the latest memory records in descending chronological order.
      tags: [Memory]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Chronological signal stream.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './schemas/MemoryRecord.schema.json'